Split out CmdStanRunner.
