+++ src/test/scala/com/cibo/scalastan/StanResultsSpec.scala
@@ -16,7 +16,7 @@ class StanResultsSpec extends ScalaStanBaseSpec {
     }
   }
 
-  private val model = CmdStanCompiledModel(new File("."), TestScalaStan, TestScalaStan.model)
+  private val model = CmdStanCompiledModel(new File("."), TestScalaStan.model)
 
 
   private val testData1 = Map[String, Int](
